Membuat class baru
Membuat class baru yang bernama Wife dan Person
