+++ Wife.java
@@ -1,0 +1,12 @@
+package lib;
+
+public class Wife extends Person{
+
+    private String spouseName;
+	private String spouseIdNumber;
+
+   	public void setSpouse(String spouseName, String spouseIdNumber) {
+		this.spouseName = spouseName;
+		this.spouseIdNumber = idNumber;
+	}
+}
